Feat: OCP 예제 코드
RoomService에서 RoomRepository를 변경해서 사용하더라도 Service코드는 변화가 없어야 한다. 이를위해 인터페이스를 생성하고 이를 구현하는 Repository 구현체를 만드는 방식으로 작성

+++ src/main/java/com/example/summer_comming/room/RoomRepository.java
@@ -2,6 +2,7 @@ package com.example.summer_comming.room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+// 껍데기다. 오버라이드 되는애를 호출함
 @Repository
-public interface  RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
 }
